fix recursive dependencies problem
