+++ packages/client/package.json
@@ -65,7 +65,6 @@
     "@orpc/shared": "workspace:*"
   },
   "devDependencies": {
-    "@orpc/openapi": "workspace:*",
     "zod": "^3.24.1"
   }
 }
