+++ packages/client/src/client.ts
@@ -1,7 +1,6 @@
-import type { Client, ClientContext } from '@orpc/client'
 import type { AnyContractRouter, ContractRouterClient } from '@orpc/contract'
 import type { AnyRouter, RouterClient } from '@orpc/server'
-import type { ClientLink } from './types'
+import type { Client, ClientContext, ClientLink } from './types'
 
 export interface createORPCClientOptions {
   /**
