+++ packages/contract/src/event-iterator.ts
@@ -1,54 +1,8 @@
 import type { StandardSchemaV1 } from '@standard-schema/spec'
 import type { Schema } from './schema'
-import { ORPCError } from '@orpc/client'
-import { getEventMeta, isAsyncIteratorObject, isEventMetaContainer, withEventMeta } from '@orpc/server-standard'
+import { mapEventIterator, ORPCError } from '@orpc/client'
+import { isAsyncIteratorObject } from '@orpc/server-standard'
 import { ValidationError } from './error'
-
-export function mapEventIterator<TYield, TReturn, TNext, TMap = TYield | TReturn>(
-  iterator: AsyncIterator<TYield, TReturn, TNext>,
-  maps: {
-    value: (value: NoInfer<TYield | TReturn>, done: boolean | undefined) => Promise<TMap>
-    error: (error: unknown) => Promise<unknown>
-  },
-): AsyncGenerator<TMap, TMap, TNext> {
-  return (async function* () {
-    try {
-      while (true) {
-        const { done, value } = await iterator.next()
-
-        let mappedValue = await maps.value(value, done) as any
-
-        if (mappedValue !== value) {
-          const meta = getEventMeta(value)
-          if (meta && isEventMetaContainer(mappedValue)) {
-            mappedValue = withEventMeta(mappedValue, meta)
-          }
-        }
-
-        if (done) {
-          return mappedValue
-        }
-
-        yield mappedValue
-      }
-    }
-    catch (error) {
-      let mappedError = await maps.error(error)
-
-      if (mappedError !== error) {
-        const meta = getEventMeta(error)
-        if (meta && isEventMetaContainer(mappedError)) {
-          mappedError = withEventMeta(mappedError, meta)
-        }
-      }
-
-      throw mappedError
-    }
-    finally {
-      await iterator.return?.()
-    }
-  })()
-}
 
 const EVENT_ITERATOR_SCHEMA_SYMBOL = Symbol('ORPC_EVENT_ITERATOR_SCHEMA')
